re-organise class hierarchy to make the code easier to reason about.
Event -> Action -> Draw -> Screen

Setup of curses environment and implementation of actions agnostic to the data
being displayed in sub-windows (such as resizing the terminal), goes in Screen.

Draw inherits from Screen and defines how to render actions related to the data
displayed in the screen's windows.

Action inherits from Draw and contains methods which define what to draw (rather
than how - Draw takes care of that.)

Event is our starting point, it contains which action to execute when an event -
such as a keypress or window resize - is instigated. It also defines the main
curses event loop, that is the entry point for our program.